JWT authorisation complete

+++ controllers/api/v1/posts_api.js
@@ -23,29 +23,21 @@ module.exports.destroy = async function (req, res) {
     try {
         let post = await Post.findById(req.params.id);
 
-        // if (post.user == req.user.id) {
+        if (post.user == req.user.id) {
             post.remove();
 
             await Comment.deleteMany({post: req.params.id});
 
-            // if (req.xhr) {
-            //     return res.status(200).json({
-            //         data: {
-            //             post_id: req.params.id
-            //         },
-            //         message: 'Post deleted'
-            //     });
-            // }
-
-            // req.flash('success', 'Post deleted!');
             return res.json(200, {
                 message: 'Post and related comment deleted successfully'
             })
-        // }else{
-        //     req.flash('error', ' You cannot delete post!');
-        //     return res.redirect('back');
-        // }
+        }else{
+            return res.json(401, {
+                message: 'You can not delete this post'
+            })
+        }
     } catch (err) {
+        console.log('*******', err);
         return res.json(500, {
             message: 'Internal server error'
         })
